refactor(tests): split fake repository reviews
